fix: use u64 for fees instead of f64

+++ smart-contracts/assembly/__tests__/helpers/math.spec.ts
@@ -19,7 +19,7 @@ describe('test sqrt calculations', () => {
 describe('test getFeeFromAmount', () => {
   test('getFeeFromAmount of 0.3% of 100 should be 0.3', () => {
     const inputAmount = parseMas(100);
-    const fee = f64(0.3 * 10_000); // 0.3% ===> 3000
+    const fee = u64(0.3 * 10_000); // 0.3% ===> 3000
     const result = getFeeFromAmount(inputAmount, fee);
     print('Input amount: ' + inputAmount.toString());
     print('Fee: ' + fee.toString());
@@ -29,7 +29,7 @@ describe('test getFeeFromAmount', () => {
 
   test('getFeeFromAmount of 0.3% of 10 should be 0.03', () => {
     const inputAmount = parseMas(10);
-    const fee = f64(0.3 * 10_000); // 0.3% ===> 3000
+    const fee = u64(0.3 * 10_000); // 0.3% ===> 3000
     const result = getFeeFromAmount(inputAmount, fee);
     print('Input amount: ' + inputAmount.toString());
     print('Fee: ' + fee.toString());
@@ -39,7 +39,7 @@ describe('test getFeeFromAmount', () => {
 
   test('getFeeFromAmount of 0.3% of 5 should be 0.015', () => {
     const inputAmount = parseMas(5);
-    const fee = f64(0.3 * 10_000); // 0.3% ===> 3000
+    const fee = u64(0.3 * 10_000); // 0.3% ===> 3000
     const result = getFeeFromAmount(inputAmount, fee);
     print('Input amount: ' + inputAmount.toString());
     print('Fee: ' + fee.toString());
